requirements.txt
boto>=2.6.0
jinja2>=2.6
GitPython>=0.3
Fabric>=1.5.2
Jinja2>=2.6
argparse>=1.2.1
